refactor: 移除集中管理的mock数据、将其分散到相关页面
