+++ pages/sale/sales/details.vue
@@ -1,5 +1,44 @@
 <script setup lang="ts">
-import { depositList, pics } from '~/types/test'
+// 测试数据。待替换
+const depositList: SalesSlip[] = [
+  {
+    num: 'CZ-32493974',
+    store: {
+      name: '青青草原一号店',
+      salesVolume: 100000,
+    },
+    member: {
+      nickname: '张三',
+      phone: '13800138000',
+      level: 2,
+    },
+    mainSalesman: '李四',
+    goodsInfo: '足金挂坠',
+    quantity: 55789,
+    realAmount: 25000,
+    goodsAmount: 15000,
+    discountAmount: '100',
+    oldDiscountAmount: '20',
+    integral: '32',
+    openType: '预定中',
+    createTime: '2021-11-11 11:11:11',
+    slipStatus: 2,
+  },
+]
+const pics = [
+  {
+    url: [
+      'https://fastly.jsdelivr.net/npm/@vant/assets/cat.jpeg',
+      'https://fastly.jsdelivr.net/npm/@vant/assets/cat.jpeg',
+      'https://fastly.jsdelivr.net/npm/@vant/assets/cat.jpeg',
+      'https://fastly.jsdelivr.net/npm/@vant/assets/cat.jpeg',
+      'https://fastly.jsdelivr.net/npm/@vant/assets/cat.jpeg',
+      'https://fastly.jsdelivr.net/npm/@vant/assets/cat.jpeg',
+    ],
+    data: '2021-11-11 11:11:11',
+  },
+]
+
 // 销售明细页面
 useSeoMeta({
   title: '销售明细',
@@ -22,10 +61,6 @@ const detailsCards = resolveComponent('sale-sales-details')
 // currentSelected.value = 2 时，切换到销售明细图
 const detailsChart = resolveComponent('sale-sales-chart')
 
-// 测试数据。待替换
-const saleList = depositList
-const pices = pics
-
 const handleClick = async () => {
 // 跳转到详情页
   await navigateTo('/sale/sales/order')
@@ -46,8 +81,8 @@ const handleClick = async () => {
       <!-- content -->
       <component
         :is="currentSelected === 1 ? detailsCards : detailsChart"
-        :info="saleList"
-        :pics="pices"
+        :info="depositList"
+        :pics="pics"
         @user-click="handleClick"
       />
     </div>
